% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{exampleMydf}
\alias{exampleMydf}
\title{\emph{piggyBac} mutant classification-data from pooled 1k heat shock screen}
\format{
A 2-column data frame with 752 \emph{P. falciparum} geneIDs in column 1, and interest-category classifications (here, phenotypes in pooled screening) in column 2.
}
\source{
\url{https://www.nature.com/articles/s41467-021-24814-1}
}
\usage{
data(exampleMydf)
}
\description{
A data frame with \emph{P. falciparum} geneIDs in column 1, and interest-category classifications in column 2. In pooled \emph{piggyBac}-mutant screening, common interest-category classifications may be some form of "sensitive", "tolerant", and "neutral".

All mutants detected above threshold in the pooled 1k-library heat shock screen are included in this table (n = 752). Derived from \url{https://www.nature.com/articles/s41467-021-24814-1}, Table S2.

can be used as example mydf input for run.topGO.meta.
}
\examples{
run.topGO.meta(mydf = exampleMydf, geneID2GO = Pfal_geneID2GO)
}
\keyword{datasets}
