% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatGOdb.curated.R
\name{formatGOdb.curated}
\alias{formatGOdb.curated}
\title{Creates new curated GO database from .gaf file retrieved from PlasmoDB.}
\usage{
formatGOdb.curated(url, organism = "PF3D7", plasmoDB_version = "")
}
\arguments{
\item{url}{url or filepath to a .gaf or gz-compressed gaf file. This should be the complete GO.gaf file (not one that has "Curated" in the PlasmoDB filename).}

\item{organism}{optional string to include in your output filename. Defaults to "PF3D7".}

\item{plasmoDB_version}{optional string to include the PlasmoDB release version in your output filename (e.g. "v68").}
}
\description{
Fetches \emph{P. falciparum} annotations having only curated evidence codes from PlasmoDB, from which it creates a new GO database compatible with topGO. The output can be used as the \code{geneID2GO} parameter for \code{\link[=run.topGO.meta]{run.topGO.meta()}}. A generally useful tool for keeping GO analyses up-to-date.
}
\details{
Outputs a dated file named "*_curated_GOdb.out" to the ./Routput folder. ./Routput will be created if it doesn't already exist.

You will need to run \code{topGO::readmappings()} on the file generated with the formatGOdb functions to use it as the geneID2GO parameter.
}
\section{\strong{Notes on gaf.gz format}}{
The .gaf or .gaf.gz file should be in tabular format with 17 columns, one row for each GO term associated with a geneID. No formatting is necessary when using the provided url.

Note that zip archives (extension: .zip) won't work directly with this function. Zip files need to be downloaded and unzipped, then the 'url' argument should be updated to the local unzipped file path.
}

\section{Notes on included GO databases}{

A GOdb from Plasmo DB's \emph{P. falciparum} annotation (from PlasmoDB beta release 69; accessed April 2026) pre-formatted using this function and ready for run.topGO.meta is included in this package (\link{Pfal_geneID2GO_curated}) and is used as the default \code{geneID2GO} input to \code{\link[=run.topGO.meta]{run.topGO.meta()}}.
}

\examples{
\dontrun{
# Generate your GO database
formatGOdb.curated(url = "https://plasmodb.org/common/downloads/release-68/Pfalciparum3D7/gaf/PlasmoDB-68_Pfalciparum3D7_GO.gaf.gz", organism = "PF3D7", plasmoDB_version = "v68")

# prepare it for pfGO
my.geneID2GO <- topGO::readmappings("./Routput/PF3D7_v68_20260422_curated_GOdb.out")

# use as input to run pipeline
run.topGO.meta(mydf, geneID2GO = my.geneID2GO)
}

}
\seealso{
Other formatGOdb functions: 
\code{\link{formatGOdb}()}
}
\concept{formatGOdb functions}
