% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatGOdb.R
\name{formatGOdb}
\alias{formatGOdb}
\title{Creates new GO database from .gaf file retrieved from PlasmoDB.}
\usage{
formatGOdb(url, organism = "PF3D7", plasmoDB_version = "")
}
\arguments{
\item{url}{url or filepath to a .gaf or gz-compressed gaf file. E.g. current latest \emph{Pf} consortium .gaf annotation file hosted at PlasmoDB: \url{https://plasmodb.org/common/downloads/release-68/Pfalciparum3D7/gaf/PlasmoDB-68_Pfalciparum3D7_GO.gaf.gz}.}

\item{organism}{optional string to include in your output file-name. Defaults to "PF3D7".}

\item{plasmoDB_version}{optional string to include the PlasmoDB release version in your output filename (e.g. "v68").}
}
\description{
Fetches \emph{P. falciparum} annotations from PlasmoDB, from which it creates a GO database compatible with topGO. The output can be used as the geneID2GO parameter for run.topGO.meta. A generally useful tool for keeping GO analyses up-to-date.
}
\section{Output}{

A dated file named "./Routput/*_GOdb.out". The ./Routput directory will be created if it doesn't already exist. Note that the PlasmoDB release version isn't recorded automatically, so it is good practice to include the current release number in the filename via the \code{plasmoDB_version} parameter for your records.

You will need to run \code{topGO::readmappings()} on the file generated with formatGOdb before using it as the geneID2GO parameter.
}

\section{Notes on .gaf format}{

The .gaf file should be in tabular format with 17 columns, one row for each GO term associated with a geneID. No formatting is necessary when using the provided url.

Note that zip archives (extension: .zip) won't work directly with this function. Zip files need to be downloaded and unzipped, then the \code{url} argument should be updated to the local unzipped file path.
}

\section{Notes on included GO databases}{

a GOdb from PlasmoDB's \emph{P. falciparum} annotation (from PlasmoDB beta release 69; accessed April 2026) pre-formatted using the this function and ready for run.topGO.meta is included in this package (\link{Pfal_geneID2GO}).
}

\section{Other notes}{

Note that this function retrieves GO annotations assigned by \strong{all} evidence-codes, including lower-confidence/automated computer-assigned annotations.

See \code{\link[=formatGOdb.curated]{formatGOdb.curated()}} for a version of this function that weeds out non-curated annotations (i.e., those with evidence codes IEA, Inferred-from-Electronic-Annotation; IBA, Inferred from Biological aspect of Ancestor; or NAS, Nontraceable Author Statements).
}

\examples{
\dontrun{
# Generate your GO database
formatGOdb(url = "https://plasmodb.org/common/downloads/release-68/Pfalciparum3D7/gaf/PlasmoDB-68_Pfalciparum3D7_GO.gaf.gz", organism = "PF3D7", plasmoDB_version = "v68")

# prepare it for pfGO
my.geneID2GO <- topGO::readmappings("./Routput/PF3D7_v68_20260422_GOdb.out")

# use as input to run pipeline
run.topGO.meta(mydf, geneID2GO = my.geneID2GO)
}

}
\seealso{
Other formatGOdb functions: 
\code{\link{formatGOdb.curated}()}
}
\concept{formatGOdb functions}
