% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{get.annot}
\alias{get.annot}
\title{get.annot}
\usage{
get.annot(x)
}
\arguments{
\item{x}{input gff file
\emph{notes on gff format}
The gff file should be in tabular format with 9 columns, one for each annotated feature associated with a geneID. No formatting is necessary when using the file at the provided url.

an annotation created from PlasmoDB's latest \emph{P. falciparum} .gff file (accessed November 1, 2021) pre-formatted using this function and ready for run.topGO.meta is included in this package (pf.annot).}
}
\description{
Extracts and formats annotations from a .gff file. Not required presently for the GO enrichment pipeline, but provides useful context for results. Can be used as-is with a provided gff file as input, or is called by get_pfannot to get the gff file from PlasmoDB.
}
\seealso{
\code{\link[=formatGOdb]{formatGOdb()}}
}
