% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{get.value}
\alias{get.value}
\title{get.value}
\usage{
get.value(id, lookupvector)
}
\arguments{
\item{id}{a character-string you want to match on}

\item{lookupvector}{a named vector you will search for your id}
}
\description{
a lookup function to match an ID and return the matching values
}
\details{
This one needs more documentation. I don't remember what it returns as I wrote it a long time ago. Could be a vector of matching values, or could be a vector of positions you can then filter the lookupvector on to return the values.
}
