% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{get_pfannot}
\alias{get_pfannot}
\title{get_pfannot}
\usage{
get_pfannot(
 
    gff_url = "https://plasmodb.org/common/downloads/Current_Release/Pfalciparum3D7/gff/data/PlasmoDB-54_Pfalciparum3D7.gff",
  organism = "Pf"
)
}
\arguments{
\item{gff_url}{connection to gff file. Defaults to "https://plasmodb.org/common/downloads/Current_Release/Pfalciparum3D7/gff/data/PlasmoDB-54_Pfalciparum3D7.gff"
\emph{notes on gff format}
The gff file should be in tabular format with 9 columns, one for each annotated feature associated with a geneID. No formatting is necessary when using the provided url.

an annotation created from PlasmoDB's latest P. falciparum gff file (accessed November 1, 2021) pre-formatted using this function and ready for run.topGO.meta is included in this package (pf.annot).}
}
\description{
Extracts and formats annotations from a gff file from plasmoDB. Not required presently for the GO enrichment pipeline, but provides useful context for results. Opens a connection to the gff file from plasmoDB without downloading it, then calls get_annot() to extract and format the annotation.
}
\seealso{
\code{\link[=get_annot]{get_annot()}}
}
