% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{insertions}
\alias{insertions}
\title{\emph{piggyBac} insertions reported in Zhang et al. 2018 (Science)}
\format{
A list of 3 dataframes--Pilot_library, corresponding to published table S1; Saturation_library, corresponding to Table S3, and GenesbyMIS, corresponding to Table S5.
}
\source{
\url{https://www.science.org/doi/10.1126/science.aap7847}
}
\usage{
data(insertions)
}
\description{
\emph{P. falciparum} insertion-data as published in Science 2018. Contains 3 of the published supplemental tables (Table S1, Pilot_library; Table S3, Saturation_library; Table S5, GenesbyMIS). Not need to run GO enrichment functions, but useful additional info to have for interpreting output.
}
\keyword{datasets}
