% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_enrichRes.R
\name{make_enrichRes}
\alias{make_enrichRes}
\title{make_enrichRes}
\usage{
make_enrichRes(
  my.cpInput,
  interestCategory,
  ontology = "BP",
  pval_cutoff = 0.05
)
}
\arguments{
\item{my.cpInput}{a reformatted df from pfGO output all.combined.sig.genes.per.sig.term.tsv (using the reformat_sigGenes function)}

\item{interestCategory}{character string of your category of interest (should match a value from the interest.category column of my.cpInput)}

\item{ontology}{one of "BP", "CC" or "MF"}

\item{pval_cutoff}{significance cutoff. Default is 0.05.}
}
\value{
an enrichResult object (which can be visualized in various ways using functions from the enrichplot package). The enrichResult object class is described in the DOSE package.
}
\description{
function to turn output of reformat_sigGenes into an enrichResult object, which can then be visualized in many different ways with functions from the enrichplot package.
}
\details{
Title
}
