% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{makeGOhierarchy.dir}
\alias{makeGOhierarchy.dir}
\title{Make "/Routput/GO/hierarchy.plots" output folder if it doesn't exist.}
\usage{
makeGOhierarchy.dir()
}
\value{
newly created path
}
\description{
This function creates the main output-directory structure I use for my topGO pipeline. All the included makeDir functions evaluate to the newly created path, or to the existing path if it already exists.
}
