% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{makeGOoutput.dir}
\alias{makeGOoutput.dir}
\title{Make "Routput/GO" directory if it doesn't already exist}
\usage{
makeGOoutput.dir()
}
\value{
newly created path
}
\description{
This function creates the main output-directory structure I use for my topGO pipeline. All the included makeDir functions evaluate to the newly created path, or to the existing path if it already exists.
}
\seealso{
Other makeDirs: 
\code{\link{makeGOhierarchy.dir}()},
\code{\link{makeGOsig.genes.dir}()},
\code{\link{makeRoutput.dir}()}
}
\concept{makeDirs}
