% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pack.test-package.R
\docType{package}
\name{pack.test-package}
\alias{pack.test}
\alias{pack.test-package}
\title{pack.test: JO's Common Functions}
\description{
This package is a test-run for functions I commonly use, particularly for \emph{P. falciparum} functional enrichment analyses.
}
\author{
\strong{Maintainer}: J Oberstaller \email{jobersta@usf.edu} (\href{https://orcid.org/0000-0002-3033-2212}{ORCID})

}
\keyword{internal}
