% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{pf.genesets.mpmp}
\alias{pf.genesets.mpmp}
\title{Curated P. falciparum gene-sets.}
\format{
A large data frame of 54090 rows and 5 columns.

\describe{
\item{Gene.ID}{P. falciparum gene ID)}
\item{Map.Name}{name of the metabolic pathway map/other gene-set curated from publication; searchable at the MPMP website)}
\item{Map_id}{the mpmp website address for the pathway map if applicable/the mpmp page for that gene-set. Follow these links to view pathways of interest.)}
\item{Map.cat}{appears to be the gene symbol (short name), or gene-product where available)}
\item{pla.id}{appears to be the GO term or enzyme-code where applicable/available)}
}
}
\source{
the Malaria Parasite Metabolic Pathways database (MPMP). \url{https://mpmp.huji.ac.il}
}
\usage{
data(pf.genesets.mpmp)
}
\description{
curated P. falciparum metabolic pathways, gene-sets, GO terms from MPMP
}
\details{
Not sure about date of access or exact origins of this dataset, but shared from Chengqi Wang in April 2020.
}
\keyword{datasets}
