% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{pf.insertions}
\alias{pf.insertions}
\title{\emph{piggyBac} insertions reported in Zhang et al. 2018 (Science)}
\format{
A list of 3 dataframes--pilot_library, corresponding to published table S1; saturation_library, corresponding to Table S3, and genesbyMIS, corresponding to Table S5. Column-names have been edited to be R-friendly.
}
\source{
\url{https://www.science.org/doi/10.1126/science.aap7847}
}
\usage{
data(pf.insertions)
}
\description{
\emph{P. falciparum} insertion-data as published in Science 2018. Contains 3 of the published supplemental tables (Table S1, pilot_library; Table S3, saturation_library; Table S5, genesbyMIS). Not needed to run GO enrichment functions, but useful additional info to have for interpreting output.
}
\keyword{datasets}
