% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfGO-package.R
\docType{package}
\name{pfGO-package}
\alias{pfGO}
\alias{pfGO-package}
\title{pfGO: Plasmodium falciparum functional enrichment analysis tools}
\description{
Handy functions and datasets specifically enabling Plasmodium falciparum functional enrichment analyses.
}
\details{
See the documentation for the \code{\link[=run.topGO.meta]{run.topGO.meta()}} function for examples and common use-cases. See the "index" link at the bottom of this window for all included datasets and functions.
}
\seealso{
\code{\link[=run.topGO.meta]{run.topGO.meta()}}
}
\author{
\strong{Maintainer}: J Oberstaller \email{jobersta@usf.edu} (\href{https://orcid.org/0000-0002-3033-2212}{ORCID})

}
\keyword{internal}
