% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reformat_sigGenes.R
\name{reformat_sigGenes}
\alias{reformat_sigGenes}
\title{Reformat sig genes output for clusterProfiler}
\usage{
reformat_sigGenes(my.sigGenes)
}
\arguments{
\item{my.sigGenes}{a dataframe of pfGO output all.combined.sig.genes.per.sig.term.tsv}
}
\description{
Reformats significant genes in significant terms output to something more compatible with clusterProfiler/enrichPlot for visualization. Output is a dataframe with the appropriate columns that can be turned into an enrichRes object (see the DOSE package), which can then be plotted with enrichPlot functions.
}
\examples{
# example code

\dontrun{
my.sigGenes <- read.delim("Routput/GO/all.combined.sig.genes.per.sig.term.tsv")
my.cp.df <- reformat_sigGenes(my.sigGenes)
}

}
\seealso{
\code{\link[=make_enrichRes]{make_enrichRes()}}
}
