% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.topGO.meta.R
\name{run.topGO.meta}
\alias{run.topGO.meta}
\title{Run topGO pipeline}
\usage{
run.topGO.meta(
  mydf = "mydf",
  geneID2GO = "Pfal_geneID2GO_curated",
  pval = 0.05,
  minTermSize = 5,
  algorithm = "weight01"
)
}
\arguments{
\item{mydf}{data frame with geneIDs in column 1, and interest-category classifications in column 2. Your input dataframe}

\item{geneID2GO}{A list of named vectors of GO IDs--one vector of GO-terms for each geneID. Defaults to the included (\link{Pfal_geneID2GO_curated}) object. See \linkSections{Using your own custom GO database} for other options.}

\item{pval}{pvalue threshold for significance. Defaults to 0.05.}

\item{minTermSize}{minimum number of genes that must be mapped to a GO term for it to be included in the enrichment analysis. Defaults to 5.}

\item{algorithm}{algorithm to run for the enrichment analysis. Accepted values are c("classic", "elim", "weight", "weight01", "lea", "parentchild"). Defaults to "weight01". See the topGO package documentation and associated publications for details on algorithms.}
}
\description{
Tests for functional enrichment in gene-categories of interest.
}
\details{
The \strong{run.topGO.meta} function:
\itemize{
\item defines which genes are "interesting" and which should be defined as background for each category specified in mydf,
\item makes the GOdata object for topGO,
\item tests each category of interest for enriched GO-terms against all the other genes included in mydf (the "gene universe"),
\item and then outputs results to table (.tsv files that can be opened in Excel).
}

Enrichments are performed by each ontology (molecular function, biological process, cellular compartment) sequentially on all groups of interest using the specified \code{algorithm}, followed by the Fisher test to assess significance (pfGO only supports the fisher test statistic at this time, as it is compatible with all supported algorithms). Results are combined in the final output table ("Routput/GO/all.combined.GO.results.tsv").

TopGO automatically accounts for genes that cannot be mapped to GO terms (or are mapped to terms with < \code{minTermSize} genes in the analysis) with "feasible genes" indicated in the topGO.log files in the "Routput/GO" folder.
}
\section{Outputs}{

run.topGO.meta creates several output-files, including:
\itemize{
\item enrichment results,
\item significant genes per significant term,
\item plots of the GO-term hierarchy relevant to the analysis, and
\item thorough log-files for each gene-category of interest tested against the background of all other genes in the analysis.
}

Primary results from run.topGO.meta will be in "Routput/GO/all.combined.GO.results.tsv".
}

\section{Concepts for common use-cases}{


\emph{RNAseq}:
In an RNAseq analysis, common categories might be "upregulated", "downregulated", and "neutral". The gene universe would consist of all genes detected above your threshold cutoffs (\emph{not necessarily all genes in the genome}).

\emph{piggyBac screens}:
In pooled \emph{piggyBac}-mutant screening, common categories might be "sensitive", "tolerant", and "neutral". The gene universe would consist of all genes represented in your screened library of mutants (\emph{again, not all genes in the genome}).
See the included \link{exampleMydf} as an example.
}

\section{Using your own custom GO database}{


A correctly formatted geneID2GO object is included for \emph{P. falciparum} enrichment analyses and is set as the default (\link{Pfal_geneID2GO_curated}). You may also provide your own, so long as it is a named character-vector of GO-terms (each vector named by geneID, with GO terms as each element).

You can use the included \code{\link[=formatGOdb.curated]{formatGOdb.curated()}} function to format a custom GO database from curated GeneDB/PlasmoDB annotations for several non-model organisms (or the \code{\link[=formatGOdb]{formatGOdb()}} function to include all GO annotations, if you aren't picky about including automated electronic annotations). If you're studying a model organism, several annotations are already available and can be downloaded through the AnnotationDbi bioconductor package.
}

\section{Note on algorithms}{

The major advantage of GO-topology-aware methods ("elim", and "weight", and the default hybrid of the two, "weight01") . . .
I find enrichment results returned via the "classic" method are much less biologically informative, as they tend to be dominated by redundant terms of the same direct parent-child lineage. But the classic method usually does return those general terms to which many genes are mapped as significant, and many times reviewers prefer to see larger numbers of significant genes to believe a term is biologically relevant. Therefore the classic method is a useful comparator for interpreting results from topology-aware methods.
}

\examples{

data(exampleMydf)  ## load exampleMydf object (mutant classification-data from pooled 1k heat shock screen)
data(Pfal_geneID2GO_curated)  ## load Pfal_geneID2GO_curated object
run.topGO.meta.multi(exampleMydf,Pfal_geneID2GO_curated)

}
\seealso{
\code{\link[topGO:topGO-package]{topGO::topGO()}}
}
