% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{formatGOdb.curated}
\alias{formatGOdb.curated}
\title{Generates new GO database from curated evidence-codes only for functional enrichment using run.topGO.meta.}
\usage{
formatGOdb.curated(
 
    url = "https://plasmodb.org/common/downloads/Current_Release/Pfalciparum3D7/gaf/PlasmoDB-56_Pfalciparum3D7_GO.gaf",
  organism = "Pf"
)
}
\arguments{
\item{url}{url to a .gaf or .gaf.gz file. Defaults to latest \emph{Pf} consortium .gaf annotation file from PlasmoDB.}

\item{organism}{optional string to include in your output file-name. Defaults to "Pf".}
}
\description{
Generates new GO database from curated evidence-codes only for functional enrichment using run.topGO.meta.
}
\details{
Outputs a dated file named "*_GOdb.out" to the ./Routput folder. ./Routput will be created if it doesn't already exist.

You will need to run \code{topGO::readmappings()} on the file generated with formatGOdb to use it as the geneID2GO parameter.

Example: geneID2GO <- topGO::readmappings("./Routput/Pf_Mar022022_GOdb.out")
}
\section{\strong{Notes on gaf.gz format}}{
The .gaf or .gaf.gz file should be in tabular format with 17 columns, one row for each GO term associated with a geneID. No formatting is necessary when using the provided url.

A GOdb from Plasmo DB's latest \emph{P. falciparum} annotation (accessed Nov 5, 2021) pre-formatted using this function and ready for run.topGO.meta is included in this package (Pfal_geneID2GO).
}

\seealso{
\code{\link[=formatGOdb]{formatGOdb()}}
}
