% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{formatGOdb.curated}
\alias{formatGOdb.curated}
\title{Generates new GO database from curated evidence-codes only for functional enrichment using run.topGO.meta.}
\usage{
formatGOdb.curated(url, organism = "Pf", plasmoDB_version = "")
}
\arguments{
\item{url}{url or filepath to a .gaf or gz-compressed gaf file. Note that zip archives won't work directly with this function. Zip files need to be downloaded and unzipped, then the 'url' argument should be updated to the local unzipped file path.}

\item{organism}{optional string to include in your output filename. Defaults to "Pf".}

\item{plasmoDB_version}{= optional string to include the PlasmoDB release version in your output filename (e.g. "v66").}
}
\description{
Generates new GO database from curated evidence-codes only for functional enrichment using run.topGO.meta.
}
\details{
Outputs a dated file named "*_GOdb.out" to the ./Routput folder. ./Routput will be created if it doesn't already exist.

You will need to run \code{topGO::readmappings()} on the file generated with formatGOdb to use it as the geneID2GO parameter.

Example: geneID2GO <- topGO::readmappings("./Routput/Pf_Mar022022_GOdb.out")
}
\section{\strong{Notes on gaf.gz format}}{
The .gaf or .gaf.gz file should be in tabular format with 17 columns, one row for each GO term associated with a geneID. No formatting is necessary when using the provided url.

A GOdb from Plasmo DB's \emph{P. falciparum} annotation (from PlasmoDB release 66; accessed Sep 28, 2023) pre-formatted using this function and ready for run.topGO.meta is included in this package (Pfal_geneID2GO).
}

\seealso{
\code{\link[=formatGOdb]{formatGOdb()}}
}
