% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{formatGOdb.curated}
\alias{formatGOdb.curated}
\title{formatGOdb.curated}
\usage{
formatGOdb.curated(
 
    gaf.gz_url = "ftp://ftp.sanger.ac.uk/pub/genedb/releases/latest/Pfalciparum/Pfalciparum.gaf.gz",
  organism = "Pf"
)
}
\arguments{
\item{gaf.gz_url}{url to a gaf.gz file. Defaults to latest Pf consortium .gaf annotation file from GeneDB.}

\item{organism}{optional string to include in your output file-name. Defaults to "Pf".}
}
\description{
Generates new GO database from curated evidence-codes only for functional enrichment using run.topGO.meta.
}
\details{
You will need to run \code{topGO::readmappings()} on the file generated with formatGOdb before using it as the geneID2GO parameter.

\emph{notes on gaf.gz format}
The gaf.gz file should be in tabular format with 17 columns, one row for each GO term associated with a geneID. No formatting is necessary when using the provided url.

a GOdb from Sanger's latest P. falciparum annotation (accessed December 8, 2020) pre-formatted using this function and ready for run.topGO.meta is included in this package (Pfal_geneID2GO).
}
