% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{formatGOdb}
\alias{formatGOdb}
\title{Creates a GO database compatible with topGO from .gaf file retrieved from PlasmoDB.}
\usage{
formatGOdb(url, organism = "Pf", plasmoDB_version = "")
}
\arguments{
\item{url}{= url or filepath to a .gaf or gz-compressed gaf file. Defaults to latest \emph{Pf} consortium .gaf annotation file hosted at PlasmoDB: \url{https://plasmodb.org/common/downloads/Current_Release/Pfalciparum3D7/gaf/PlasmoDB-CURRENT_Pfalciparum3D7_GO.gaf.gzip}. Note that zip archives won't work directly with this function. Zip files need to be downloaded and unzipped, then the 'url' argument should be updated to the local unzipped file path.}

\item{organism}{= optional string to include in your output file-name. Defaults to "Pf".}

\item{plasmoDB_version}{= optional string to include the PlasmoDB release version in your output filename (e.g. "v66").}
}
\description{
Fetches \emph{P. falciparum} annotations from PlasmoDB, from which it creates a GO database compatible with topGO. The output can be used as the geneID2GO parameter for run.topGO.meta. A generally useful tool for keeping GO analyses up-to-date.
}
\details{
Outputs a dated file named "*_GOdb.out" to the ./Routput folder. ./Routput will be created if it doesn't already exist. Note that the PlasmoDB release version isn't recorded automatically in the filename, so it is good practice to include the current PlasmoDB release number in the filename via the 'organism' parameter for your records.

You will need to run \code{topGO::readmappings()} on the file generated with formatGOdb before using it as the geneID2GO parameter.

Example: geneID2GO <- topGO::readmappings("./Routput/Pf_Mar022022_GOdb.out")
}
\section{\strong{Notes on .gaf format}}{
The .gaf file should be in tabular format with 17 columns, one row for each GO term associated with a geneID. No formatting is necessary when using the provided url.

Retrieves GO annotations assigned by \strong{all} evidence-codes. A version that weeds out any non-curated, Inferred-from-Electronic-Annotation assignments, Inferred from Biological aspect of Ancestor, or Nontraceable Author Statements is also included in this package (evidence codes IEA,IBA,NAS; see \code{\link{formatGOdb.curated}})

a GOdb from PlasmoDB's \emph{P. falciparum} annotation (from PlasmoDB release 66; accessed Sep 28, 2023) pre-formatted using the \emph{curated} version of this function and ready for run.topGO.meta is included in this package (Pfal_geneID2GO_curated).
}

\seealso{
\code{\link[=formatGOdb.curated]{formatGOdb.curated()}}
}
