% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_enrichRes.R
\name{make_enrichRes}
\alias{make_enrichRes}
\title{make_enrichRes}
\usage{
make_enrichRes(
  my.cpInput,
  interestCategory,
  ontology = "BP",
  pval_cutoff = 0.05
)
}
\arguments{
\item{my.cpInput}{a reformatted dataframe from pfGO output file all.combined.sig.genes.per.sig.term.tsv (reformatted using the \code{\link[=reformat_sigGenes]{reformat_sigGenes()}} function)}

\item{interestCategory}{character string of your category of interest (should match a value from the interest.category column of my.cpInput)}

\item{ontology}{one of "BP", "CC" or "MF"}

\item{pval_cutoff}{significance cutoff. Default is 0.05.}
}
\value{
An enrichResult object (which can be visualized in various ways using functions from the enrichplot package). The enrichResult object class is described in the DOSE package.
}
\description{
Function to turn output of \code{\link[=reformat_sigGenes]{reformat_sigGenes()}} into an enrichResult object, which can then be visualized in many different ways with functions from the enrichplot package.

\href{https://yulab-smu.top/biomedical-knowledge-mining-book/enrichplot.html}{See an enrichplot visualization overview here}.
}
\examples{
# example pfGO-output-to-visualization workflow
\dontrun{
my.sigGenes <- read.delim("Routput/GO/all.combined.sig.genes.per.sig.term.tsv")
my.cp.df <- reformat_sigGenes(my.sigGenes)
my.enrichRes <- make_enrichRes(my.cp.df, interestCategory = "sensitive", ontology = "BP")

## from this point, enrichment results can be visualized many different ways with functions included in the enrichplot package.

barplot(my.enrichRes)
cnetplot(my.enrichRes)
dotplot(my.enrichRes)

}
}
