% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pack.test-package.R
\docType{package}
\name{pack.test-package}
\alias{pack.test}
\alias{pack.test-package}
\title{pack.test: JO's Common Functions TEST}
\description{
This package is a test-run for functions I commonly use, particularly for P. falciparum functional enrichment analyses.
}
\author{
\strong{Maintainer}: J Oberstaller \email{jobersta@usf.edu} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
