% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{pf.annot}
\alias{pf.annot}
\title{All \emph{P. falciparum} gene-product annotations.}
\format{
A data frame of 5562 rows and 9 columns.

\describe{
\item{organism_name}{\emph{P. falciparum}}
\item{seqid}{chromosome ID.}
\item{source}{annotation source.}
\item{type}{type of annotation. Here only protein-coding genes and ncRNA genes are kept.}
\item{feature_start}{gene start coordinate.}
\item{feature_end}{gene end coordinate.}
\item{strand}{gene strand.}
\item{geneID}{gene ID.}
\item{description}{functional annotation}
\item{geneName}{the gene symbol, or short name, if one exists}
}
}
\source{
\url{https://plasmodb.org/common/downloads/Current_Release/Pfalciparum3D7/gff/data/PlasmoDB-54_Pfalciparum3D7.gff}
}
\usage{
data(pf.annot)
}
\description{
A dataset containing all \emph{P. falciparum} gene-product annotations taken directly from gff file. Only "gene" entries are kept to remove redundancy.
}
\details{
\subsection{Keep in mind:}{

These data aren't explicitly required for running any enrichments with this package. They are included for reference to aid exploring your enrichment results.

Some redundant columns were filtered/removed from the original data source.
Updated versions can be generated using the get_pfannot function.
}
}
\keyword{dataset}
