% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Pfal_geneID2GO}
\alias{Pfal_geneID2GO}
\title{All curated GO terms mapped to \emph{P. falciparum} genes}
\format{
A list of 3381 named vectors--one vector for each Pf geneID to which GO terms are mapped. Each vector contains all curated GO-terms mapped to the geneID.
}
\source{
\url{ftp://ftp.sanger.ac.uk/pub/genedb/releases/latest/Pfalciparum/Pfalciparum.gaf.gz}
}
\usage{
data(Pfal_geneID2GO)
}
\description{
A \emph{P. falciparum} GO database containing all curated GO terms mapped to \emph{P. falciparum} genes (accessed from GeneDB Dec 08, 2020).

can be used as the geneID2GO input for run.topGO.meta.
}
\examples{
run.topGO.meta(mydf = mydf, geneID2GO = Pfal_geneID2GO)
}
\keyword{datasets}
