% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Pfal_geneID2GO}
\alias{Pfal_geneID2GO}
\title{Pfal_geneID2GO: A curated P. falciparum GO database (accessed from GeneDB Dec 08, 2020).

A dataset containing all curated GO terms mapped to P. falciparum genes.}
\format{
A list of 3381 named vectors--one vector for each Pf geneID to which GO terms are mapped. Each vector contains all curated GO-terms mapped to the geneID.
}
\source{
\url{ftp://ftp.sanger.ac.uk/pub/genedb/releases/latest/Pfalciparum/Pfalciparum.gaf.gz}
}
\usage{
Pfal_geneID2GO
}
\description{
formatted into the geneID2GO input for run.topGO.meta from the source-url using the included format.curated.GOdb function followed by \code{topGO::readMappings()}
}
\keyword{datasets}
