% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Pfal_geneID2GO_curated}
\alias{Pfal_geneID2GO_curated}
\title{All curated GO terms mapped to \emph{P. falciparum} genes}
\format{
A list of 3454 named vectors--one vector for each \emph{Pf} geneID to which GO terms are mapped. Each vector contains all curated GO-terms mapped to the geneID.
}
\source{
\url{https://plasmodb.org/common/downloads/Current_Release/Pfalciparum3D7/gaf/PlasmoDB-66_Pfalciparum3D7_GO.gaf.gzip}
}
\usage{
data(Pfal_geneID2GO_curated)
}
\description{
A \emph{P. falciparum} GO database containing all curated GO terms mapped to \emph{P. falciparum} genes (from PlasmoDB release 66; accessed Sep 28, 2023).

Can be used as the geneID2GO input for run.topGO.meta.
}
\examples{
run.topGO.meta(mydf = exampleMydf, geneID2GO = Pfal_geneID2GO_curated)
}
\keyword{datasets}
