% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfGO-package.R
\docType{package}
\name{pfGO-package}
\alias{pfGO}
\alias{pfGO-package}
\title{pfGO: \emph{P. falciparum} functional enrichment analysis tools}
\description{
Handy functions and datasets specifically enabling \emph{Plasmodium falciparum} functional enrichment analyses.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/oberstal/pfGO}
  \item Report bugs at \url{https://github.com/oberstal/pfGO/issues}
}

}
\author{
\strong{Maintainer}: J Oberstaller \email{jobersta@usf.edu} (\href{https://orcid.org/0000-0002-3033-2212}{ORCID})

}
\keyword{internal}
