% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{run.topGO.meta}
\alias{run.topGO.meta}
\title{Run JO's topGO pipeline.}
\usage{
run.topGO.meta(mydf = "mydf", geneID2GO = "geneID2GO")
}
\arguments{
\item{geneID2GO}{a data frame of 2 columns, with geneIDs in column 1, and comma-separated GOterms in column2}

\item{data}{frame with geneIDs in column 1, and interest-category classifications in column 2}
}
\value{
writes enrichment results, sig genes per term, plots of the GO-term hierarchy, and thorough log-files for each gene-category of interest tested against the background of all other genes in the analysis.
}
\description{
This function tests for functional enrichment in gene-categories of interest.
}
